Renamed LAS files processing scripts
