+++ LASfiles_preprocessing_exec.py
@@ -1,5 +1,5 @@
 
-import LASfileTimeExtractor
+import LASfiles_preprocessing
 import time
 
 start_time = time.time()
@@ -10,7 +10,7 @@ merged_folder = r"D:\NRCAN2\FloodToolsOpenGIS\Test_LAS_GR\merged"
 UTC = -4
 
 # First tool: Filter las or laz files by days of LiDAR acquisition
-LASfileTimeExtractor.execute_extract_bydays(input_laz_folder, UTC, bydays_folder)
+LASfiles_preprocessing.execute_extract_bydays(input_laz_folder, UTC, bydays_folder)
 # Second tool: Ground points classification and filtering
 LASfileTimeExtractor.execute_groundclassification(bydays_folder, ground_folder)
 # Third tool: Merging together tiles of same day of LiDAR acquisition
